refactor(strings): Change the package name and move the `UUID()`` to the `strings` package
