--- generator/uuid/uuid.go
@@ -1,8 +1,0 @@
-package uuid
-
-import "github.com/google/uuid"
-
-// Generate uuid
-func Generate() string {
-	return uuid.New().String()
-}
